+++ src/com/dci/intellij/dbn/object/common/sorting/DBMethodNameComparator.java
@@ -25,8 +25,8 @@ public abstract class DBMethodNameComparator<T extends DBMethod> extends DBObjec
     }
 
     private static int compareByName(DBMethod method1, DBMethod method2) {
-        int overload1 = method1.getOverload();
-        int overload2 = method2.getOverload();
+        short overload1 = method1.getOverload();
+        short overload2 = method2.getOverload();
         int nameComparison = method1.getName().compareTo(method2.getName());
         if (nameComparison == 0) {
             return overload1 - overload2;
